refactor(4710): extract storage/maintenance wrapper jobs from maintenance.rs giant
